Update 2018.05.24 * run 하면 결과가 return되는 구조로 변경 * 필요없는 모듈 import제거 * int -> float 으로 변경

+++ tensorflow_model/testing_mc_static.py
@@ -37,7 +37,7 @@ def mc_run( feature_vector ) :
         output = np.array(sess.run(y_test, feed_dict={x: [feature_vector]})).reshape([-1])
         group = []
         for i in range(OUTPUT_SIZE) :
-            group.append( int(output[i] * 100) )
+            group.append(output[i] * 100)
     return group
 
 def load_data( file_path ) :
